// Clients/src/presentation/structures/AssessmentTracker/record-keeping.subtopic.ts
export const RecordKeeping = [
  {
    id: 1,
    title: "AI model capability assessment",
    questions: [
      {
        id: 1,
        question:
          "What performance criteria have been established for the AI application? How were these criteria determined, and how do they relate to the application's objectives?",
        hint: "Explain how performance criteria were determined for the AI system and how they align with the application's objectives.",
        priorityLevel: "medium priority",
        answerType: "Long text",
        inputType: "Tiptap area",
        evidenceFile: "Not required",
      },
      {
        id: 2,
        question:
          "Describe the policies and procedures in place for retaining automatically generated logs for a minimum of six months to ensure adequate data tracking and auditability.",
        hint: "Describe policies and procedures for retaining logs for a minimum of six months to ensure data tracking and auditability.",
        priorityLevel: "medium priority",
        answerType: "Long text",
        inputType: "Tiptap area",
        evidenceFile: "Not required",
      },
      {
        id: 3,
        question:
          "How has your organization tested the model's performance on extreme values and protected attributes? What were the results of these tests, and how have they informed further development?",
        hint: "Detail your organization's tests on extreme values and protected attributes, their results, and how these informed further development.",
        priorityLevel: "medium priority",
        answerType: "Long text",
        inputType: "Tiptap area",
        evidenceFile: "Not required",
      },
      {
        id: 4,
        question:
          "What patterns of failure have been identified in the model? How were these patterns discovered, and what do they reveal about the model's limitations?",
        hint: "Desribe the identified failure patterns in the model, how they were discovered, and what they reveal about the system's limitations.",
        priorityLevel: "medium priority",
        answerType: "Long text",
        inputType: "Tiptap area",
        evidenceFile: "Not required",
      },
    ],
  },
];